improve limit select style and position
